Move pupil model creation code from example to package
That way it is accessible to end users.
The example now uses the packaged functions
